% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artistic.R
\docType{data}
\name{artistic}
\alias{artistic}
\title{Artistic palettes}
\format{An object of class \code{list} of length 5.}
\usage{
artistic
}
\description{
This selection of color palettes contains a list of 5 palettes with 10
colors each that act as being aesthetically pleasing.
}
\details{
The available palettes are:
dreaming
jojo
beach
waterfall
sunset
}
\keyword{datasets}
