% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfws.R
\docType{data}
\name{usfws}
\alias{usfws}
\title{usfws palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
usfws
}
\description{
This selection of color palettes contains a list of palettes derived from this
U.S. Fish and Wildlife Services [Instagram post](https://www.instagram.com/p/CFP2H7JBSxR/?igshid=13sg433qx7ts6).
}
\details{
The available palettes are:
sockeye_salmon
brook_trout
dolly_varden
florida_gar
}
\keyword{datasets}
